LayoutTests/inspector/extensions/extensions-expected.txt

Tests WebInspector extension API

Started extension.
Running tests...
RUNNING TEST: extension_testCreatePanel
done createPanel
Panel created
{
    onSearch : {
        addListener : <function>
        removeListener : <function>
    }
}
RUNNING TEST: extension_testCreateSidebar
Sidebar created
{
    setHeight : <function>
}
RUNNING TEST: extension_testExtensionWatchSidebarExpression
Sidebar content: titlef0: "expression"f1: undefinedf2: nullf3: Objectf4: Array[0]f5: Array[3]f6: Objectf7: 42f8: Locationf9: HTMLCollection[2]f10: function () {}__proto__: Object
RUNNING TEST: extension_testExtensionWatchSidebarObject
Watch sidebar created, callback arguments dump follows:
{
    0 : {
        onUpdated : {
            addListener : <function>
            removeListener : <function>
        }
        setExpression : <function>
        setObject : <function>
        setHeight : <function>
    }
}
Sidebar content: f0: "object"f1: undefinedf2: nullf3: {}f4: []f5: [aa, bb, cc]f6: {f60:42, f61:foo, f62:[]}f7: 42
All tests done.
